+++ src/components/_atoms/buttons/navigation/ListNavigationButtonNext.tsx
@@ -1,6 +1,6 @@
 import { LinkProps } from 'react-router';
-import { NavigationIconButton } from '../../../../pages/tasks/nvi/nvi-candidate-page/_components/NavigationIconButton';
 import { dataTestId as testIds } from '../../../../utils/dataTestIds';
+import { NavigationIconButton } from '../NavigationIconButton';
 
 type ListNavigationButtonNextProps = Pick<LinkProps, 'to' | 'state'> & { title: string; dataTestId?: string };
